# Conflicts:
#	build.gradle
#	src/main/kotlin/me/steven/indrev/IndustrialRevolution.kt
#	src/main/kotlin/me/steven/indrev/IndustrialRevolutionClient.kt
#	src/main/kotlin/me/steven/indrev/blockentities/MachineBlockEntity.kt
#	src/main/kotlin/me/steven/indrev/blockentities/MultiblockBlockEntityRenderer.kt
#	src/main/kotlin/me/steven/indrev/blockentities/storage/TankBlockEntity.kt
#	src/main/kotlin/me/steven/indrev/components/TemperatureComponent.kt
#	src/main/kotlin/me/steven/indrev/components/fluid/FluidComponent.kt
#	src/main/kotlin/me/steven/indrev/components/multiblock/MultiBlockComponent.kt
#	src/main/kotlin/me/steven/indrev/items/misc/IRWrenchItem.kt
#	src/main/kotlin/me/steven/indrev/registry/IRBlockRegistry.kt

+++ src/main/kotlin/me/steven/indrev/gui/widgets/misc/WText.kt
@@ -19,13 +19,13 @@ class WText(
     )
 
     override fun paint(matrices: MatrixStack?, x: Int, y: Int, mouseX: Int, mouseY: Int) {
-        var alignedX = x
+        var alignedX = x.toDouble()
         val text = string()
         if (alignment == HorizontalAlignment.CENTER) {
-            alignedX -= MinecraftClient.getInstance().textRenderer.getWidth(text) / 2
+            alignedX -= MinecraftClient.getInstance().textRenderer.getWidth(text) / 2.0
             MinecraftClient.getInstance().textRenderer.draw(matrices, text, alignedX.toFloat(), y.toFloat(), color)
             return
         }
-        ScreenDrawing.drawString(matrices, text.asOrderedText(), alignment, alignedX, y, this.width, color)
+        ScreenDrawing.drawString(matrices, text.asOrderedText(), alignment, alignedX.toInt(), y, this.width, color)
     }
 }
